refactor: Move `getPrivateMessages` to `messageSelectors`.
Before it was in `baseSelectors`, which is not a right place for it.
All selectors related to `message` are present in `messageSelectors`,
so it's better to move it.
